Chaos Cloth Bending Element constraint fixes.
There were a few bugs with the bending element constraints -- firstly since it was using acos to calculate the bend angle, it could only calculate an angle between 0 and pi rather than -pi to pi. Switched to use atan2 instead.
Secondly, there were bugs in the gradient calculation. I'm not quite sure where it came from, but numerical differentiation tests showed it was not correct, and the constraints were not stable in practice.

Updated with new gradient calculations (based on atan2). I did numerical differentiation tests, and they now pass. I also added a limit to how large of a step you can take in a single iteration (pi/4) since the linearization (gradients) of the constraints will not produce a good direction for taking larger steps.

In testing, the constraints are now stable, assuming you have stretch and or area constraints to resist large stretching deformations. These constraints are also much better than the spring-based  bending constraints at holding nearly flat configurations.

It should be noted that this code has not been optimized to use parallelism or ispc, so they are very slow. Optimization will be done in a later pass once additional improvements to the behavior of this constraint are made (namely biphasic stiffness as well as warp vs weft directional stiffnesses).

#rb kriss.gossart
#preflight 625f002b034d8924cdd01f01

[CL 19814127 by Alex McAdams in ue5-main branch]