ConcertSyncClient: use FSoftObjectPath::ResolveObject() when finding object outers in ConcertSyncClientUtil::GetObject()
In WorldPartition levels, the path of the package to which an object is outered can vary
depending on the runtime context, e.g. editor vs. PIE vs. -game, etc. When a transaction
that affects an object is recorded on one client in a Multi-User session (for example, an editor),
the path to that object may not be the same on another client (for example, an nDisplay node
which runs using -game). This can manifest as object changes on one client not having an
effect on other clients in the session.

With this change, we use FSoftObjectPath's object resolution to find the object referenced
by the path rather than looking specifically for an object with that path using StaticFindObject().
This ensures that we pass through the ULevel's subobject resolution, in which there's a custom
resolver for subobjects inside a UWorldPartition.

Note that this provides a solution for resolving an unpartitioned path to a partitioned path.
A different solution would be needed for the reverse direction, to get an unpartitioned path from
a partitioned one, but that's not currently a situation that arises in Multi-User.

#jira UE-144253, UE-143928
#rb richard.malo, francis.hurteau, jason.walter
#preflight 625de2aeb21bb49791d019ac

[CL 19832372 by matt johnson in ue5-main branch]