Sequencer: Added a new option, "Swap Root Bone With Component Root", to put the root motion on the skeletal mesh component.
When this option is enabled, the root bone is evaluated as usual, cached and then set to 0. The cached root bone transform is then set on the skeletal mesh component in FAnimSequencerInstanceProxy::PostEvaluate. If there is an existing transform track on the skeletal mesh component, however, the cached root bone transform is applied additively to that. Without this, the skeletal mesh wouldn't be able to be globally transformed by Sequencer.

Deprecated UpdateAnimTrackWithRootMotion so that it now takes FAnimSequencerData instead of each individual parameter separately.

#jira UE-113072
#preflight 6261a7a67d162829ae87670e
#rb mike.zyracki

[CL 19872424 by Max Chen in ue5-main branch]