EditorPerf: Require statistics mentioned in EditorPerfStats.csv (allowlist) to exist in the data
### Features

Prior to this change, EditorPerf can sliently fail when statistics in the csv aren't found in the data; causing the telemetry data to be missing. This could be because asset or configuration changes modified the test routine. This change causes this missing data to print errors and fail the commandlet. This makes maintenance easier because it prevents likely incorrect data from being pushed to telemetry.

Add -alltelemetry flag for unfiltered telemetry. This flag is for scale testing and local development purproses.

### Testing

Verified to work properly with invalid statistics in the csv, and corrected statistics in the csv
Expected results running with and without -alltelemetry (see SummarizeArgs in the build graph)

### Tags

#rnx
#jira none
#rb patrick.laflamme
#preflight 626a003af97c319bebd7d2aa

[CL 19961093 by geoff evans in ue5-main branch]