Logging: Flush logs when an assertion fails
The previous call to FlushThreadedLogs serialized buffered logs to the output devices, but did not flush the output devices. Recent changes to logging mean that Flush is now thread-safe and can be called here to flush the output devices too. This means that the log file, console, and other output devices will contain every log prior to the failed assertion, and will be visible when we break in the debugger.

#preflight 626c00e1ce3959ce8fe280fe
#rb Zousar.Shaker
#fyi Rob.Krajcarski

[CL 19980264 by Devin Doucette in ue5-main branch]