Add Lumen support for multi-layer Strata material.
This CL changes LumenScreenProbeGather and LumenReflection to support 'overflow' tiles which allows to lit multiple layers per pixels. The screen is now divided into the 'primary' space (i.e., the regular view rect), and an 'overflow' space (offset at the bottom of the view texture, and whose size is configurable with a CVars). During the classification/integration/... passes, an extra indirect pass is done for computing the overflow tiles. This 'flatten' layout also allows to bookeep history data from frame to frame.

There are a number of follow up to this CL when Strata is enabled:
* Multi-view code is in place but not working correctly yet when Strata is enabled
* Reflection bilateral filter place is not fully stable yet
* UAV overlap for resource written by the regular & overflow pass are not put in place yet
* 'Overflow' tile can have border filtering artifact that needs to be addressed.

Performance and occupancy regression have done on PS5 on CitySample.

#rb daniel.wright, sebastien.hillaire
#preflight 624ed58c9660409d77a83309

#ROBOMERGE-AUTHOR: charles.derousiers
#ROBOMERGE-SOURCE: CL 20008236 via CL 20008250 via CL 20008258
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v943-19904690)

[CL 20009992 by charles derousiers in ue5-main branch]