Scripted Asset Actions now work with collection parameters.
CPF_Parm was not being applied to child properties (ie. items in the array), and the IsPropertyVisible delegate wasn't checking parent properties for the flag, which was causing only the root to be visible. This is a result of a change made to the details view to run the IsPropertyVisible delegate more frequently - previously it was only run when initializing the property tree.

#jira UE-148027
#review-20008735 @jurre.debaare, @lauren.barnes
#preflight 6270dd90ae03caabf7103e9e

[CL 20023534 by sebastian nordgren in ue5-main branch]