UE-149122 - TMap Values in User Structs do not convert to real numbers
The underlying cause is connected to inconsistent usage of serialization methods among BP-related types.

Both FEdGraphPinType and FStructVariableDescription have a FEdGraphTerminalType field type. FEdGraphTerminalType will only perform versioning fixup in operator<<. This implies that fixup can only be achieved when an outer type uses manual serialization and calls << on the type, which FEdGraphPinType does. However, FStructVariableDescription does *not* have a Serialize function, and expects tagged property serialization to fixup everything in its FEdGraphTerminalType member, which never happens.

Since this is needed as a hotfix, the easiest fix is to update FStructVariableDescription::PostSerialize to perform the fixup for FEdGraphTerminalType, which is what FEdGraphTerminalType already does in operator<<.

#jira UE-149122
#preflight 6271561f9d6c2f8f5b234a34
#rb marc.audy

#ROBOMERGE-AUTHOR: dave.jones2
#ROBOMERGE-SOURCE: CL 20027908 in //UE5/Release-5.0/... via CL 20028986
#ROBOMERGE-BOT: UE5 (Release-Engine-Staging -> Main) (v943-19904690)

[CL 20035000 by dave jones2 in ue5-main branch]