Chaos Cloth: Update PBDStiffness so that exponential "ParameterFit" is 0 when Value = 0 and 1 when Value = 1.
This is just a slight scaling and translation from the original ParameterFit function. (Previous range was [1e-3, 1] instead of  [0,1]). There is no upgrade between old and new function because the difference largest at InValue = 0 (where ParameterFitOld = 1e-3 and ParameterFitNew = 0), and behavior differences are small at this end because the curve is so flat (with the exception of seeing a difference between true 0 and anything > 0).  I did some toggle tests between the two functions, and they didn't make a perceptable difference. I recognize this also means that maybe this change is pointless, but it makes the exponential stiffness value smooth rather than having a small discontinuity near zero with the way we clip values which feels like a good thing in general. (New function is 7e-11 at the clipping point vs 1e-3).

Also, I noticed in testing that when I added the Buckling, it caused BendingSprings to be created when BendingStiffness == 0 since BucklingStiffness (unused by BendingSprings) is defaulted to 1. This has been fixed in this CL.

Sidenote: XPBDStiffness suffers from similar issues of 0 not mapping to 0. However, we actually always use Alpha = 1 / (Stiffness * Dt * Dt), so additional changes are necessary to handle true 0. I am not making those changes now because we'd like to rework XPBDStiffness units soon and XPBD is still not actually exposed in the UI.

#preflight 6272c1592f6d177be3d3cbda
#rb kriss.gossart

[CL 20085954 by Alex McAdams in ue5-main branch]