Chaos Cloth: switch XPBD constraints to use traditional stiffness inputs rather than PBD exponential compliance inputs (PBDStiffness).  We might ultimately want to have a UI that supports both an exponential 0-1 input and the actual stiffness values, but for now, using the stiffness values. With the new stiffness handling, it is possible to get stiffness = 0 (as opposed to a very small number). Update XPBD code to handle this case since we do calculations in terms of alpha = 1 / (stiffness * dt *dt).
Add XPBDBendingConstraints. In adding the XPBDBendingConstraints, I realized that I had misused VectorSelect instead of just the builtin ispc select function for scalars. This caused weird bugs in XPBDBending. I went back and fixed PBDBendingConstraints.ispc as well.

The PBDBendingConstraints change is the only change that affects existing standard users because XPBD continues to be unavailable in the UI.

#preflight 627957a8dd213363db77340b
#rb kriss.gossart

[CL 20107568 by Alex McAdams in ue5-main branch]