PThread: Added error checking to thread joining in Kill and WaitForCompletion.
Internally, we call pthread_join in these function, which may actually fail. Some of the reasons are:
   - A thread trying to join itself.
   - A deadlock e.g. two threads trying to join each other.
   - Another thread is already joining the thread.

If one of this cases happens, the following snippet would likely crash, because we used to unconditionally return 'true':

if (Thread->Kill(true))
    delete Thread;

In many places we don't even perform this check but just kill and delete threads. For this reason, I've also added a log message to provide some details if we ever crash because of this again. An assert could be a better way of dealing with it, but because Kill (public API) returns true/false to indicate if the thread has been really stopped, we can't assume that no one has code relying on it.

#jira UE-152516
#rb jeff.newquist, Jens.Petersam
#preflight 628676475a69ba071e316e9f

[CL 20296849 by Wojciech Krywult in ue5-main branch]