PoseSearch: Extensible channels Schemas are now composed of user defined channels. Channels are responsible for describing their features to the schema, indexing assets, query construction, and debug drawing. Existing schemas are migrated to the default pose or trajectory channel implementations
This change will cause some minor data loss. Schemas will automatically migrate, but database default weights and per-group weights will have to be setup again. The weights structure is now composed of an array of channel weight params rather than hard-coded pose and trajectory weights. The number of entries in the weights array is expected to mirror the number of channels setup on a Schema, but there is no UI feedback about that yet and the weights system will fallback to defaults when there aren't sufficient entries in the array.

#rb cesar.castro,samuele.rigamonti
#preflight 628fe0aee73f637548edf52d

[CL 20385207 by braeden shosa in ue5-main branch]