ParallelFor: Cancel unused tasks and time the wait for tasks to complete
When worker threads are heavily utilized by slow tasks, ParallelFor may execute most of the work on the calling thread before most (or any) of its tasks have executed. Allowing these tasks to execute later is wasteful because each task executed on a background worker pays about 0.3us to lower and then raise thread priority. Canceling the tasks can avoid that cost when paired with an upcoming change to the scheduler, which saves about 40us for each call to ParallelFor in a heavily-utilized pool of 126 worker threads.

#preflight 62587f45010ebc5d4e89d347
#rb Andriy.Tylychko, Arne.Schober

[CL 20392225 by Devin Doucette in ue5-main branch]