PR #8680: [SQLiteCore] Fix compile error when SQLITE_DEBUG enabled (Contributed by dorgonman)
The mentioned error is a clang error. It does not repro in cl.exe. Reading over this explanation of temp copies and constructor choice: https://stackoverflow.com/questions/21708606/why-does-an-in-place-member-initialization-use-a-copy-constructor-in-c11, initializer list syntax seems like a resonable solution. Delegating a call to the TAtomic<> ctor in the struct ctor init list would just add more #ifdef and hurt reability. Accepting.


#rnx
#rb geoff.evans
#jira UE-136380
#preflight 62911be7532586228d89e6a0

[CL 20397181 by dorgonman in ue5-main branch]