[TraceLog] Add OnConnection callback
Adds a callback which is issued from Trace whenever a new connection is established. Any event emitted inside the callback are collected separately and written to the trace stream after important events, but before events collected in the tail buffer. In the engine layer the callback is exposed as static delegate on TraceAuxiliary.

#rb martin.ridgers, ionut.matasaru
#preflight 6298bbad63529f84faeab10b

[CL 20468957 by Johan Berg in ue5-main branch]