EditorPerf: Memory, Rate, and better Counter statistics
### Features

Instrumentation is added to the FGenericStatsUpdater to push the sampled memory stats to counters. Ideally we would add TraceStat support for EditorPerf, but that is significantly more work.

CountPerSecond statistic has been added to compute rate information from the first start and last end timestamps of each Scope or Counter.

This change adds full min, max, mean, deviation statistics to Counters. It's all made available similarly to Scopes.

### Notes

Welford's mean, variance, and deviation are moved out FSummarizeScope to static utility functions.

This change closes a hole in FSummarizeScope::Merge() where Mean and Variance were not being handled at all. The variance merge math is probably not accurate, but it should yield something proportional. A follow up change will address that.

This change tidies up some class names and code formatting, and grouping the Scope, Counter, and Bookmark classes to be adjacent in the cpp file to make reading easier.

#rnx
#rb patrick.laflamme
#preflight 6298d10b1941378d7c6a0c63
#jira UE-154789, UE-150298

[CL 20473282 by geoff evans in ue5-main branch]