Scene Outliner: Remove forcibly showing actors with the SequencerActor tag. Spawnables are still shown in the World Outliner still because they are RF_Transient and the World Outliner is also by default set to show transient actors.
This fixes an issue where in an actor picker menu, you could pick a spawnable actor and it's not desirable to make that reference to a transient actor.

If you need to reference a spawnable from sequencer, you need to make an explicit actor reference property track and choose the actor from the picker in the track. You will no longer be able to pick a spwanble from the details panel for genuine operations such as autokeying a property that already has an actor reference track. You will need to choose it from the picker in the track.

#jira UE-105422
#preflight 6291afc6bc1213abd7fc579b
#rb andrew.rodham

[CL 20487934 by Max Chen in ue5-main branch]