Trace+Insights: Reduce header inclusion in AnalysisService.h
### Features

No functional changes. AnalysisService is only really used in a handful of places, it includes a bunch of model classes, and is generally costly to understand (and probably compile).

### Notes

This change is preliminary for a future change that adds new includes that would be circular without this change as a preliminary change. This change is also preliminary to simplify reviewing the future functional change.

#rnx
#rb ionut.matasaru
#jira UE-120810
#preflight 62a13409e820a589df6189d7

[CL 20575150 by geoff evans in ue5-main branch]